requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -p 3.11 --no-cache --no-strip-extras requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   starlette
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
distro==1.9.0
    # via openai
fastapi==0.113.0
    # via -r requirements.in
google-api-core[grpc]==2.19.2
    # via google-cloud-secret-manager
google-auth==2.34.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
google-cloud-secret-manager==2.20.2
    # via -r requirements.in
googleapis-common-protos[grpc]==1.65.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-secret-manager
grpcio==1.66.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.66.1
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   openai
    #   python-telegram-bot
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.5.0
    # via openai
openai==1.43.0
    # via -r requirements.in
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==5.28.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.8.2
    # via
    #   fastapi
    #   openai
pydantic-core==2.20.1
    # via pydantic
python-telegram-bot[webhooks]==21.6
    # via -r requirements.in
pytz==2024.2
    # via -r requirements.in
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   tiktoken
    #   youtube-transcript-api
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.34
    # via -r requirements.in
starlette==0.38.4
    # via fastapi
tenacity==9.0.0
    # via -r requirements.in
tiktoken==0.7.0
    # via -r requirements.in
tornado==6.4.1
    # via python-telegram-bot
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.2
    # via requests
uvicorn==0.30.6
    # via -r requirements.in
youtube-transcript-api==0.6.2
    # via -r requirements.in
